Add files via upload
I had a pretty hard time making all the curves in the photo as well as all the dots in the circle, that's why the dots are not very accurately duplicated.

BUT, after a while, I did finally figure out how to makes multiple dots using the fore loop function we learned the other day as well as learning how to use the "curve()" function but I'm still a little confused about inserting control points to control the amount of that the line is curved. Also, is there a simpler way to create the grid of dots and create a boundary so that the dots are only drawn within the circle?